Feature/deal/migrations

+++ Deal/src/main/java/org/deal/model/Employment.java
@@ -1,10 +1,12 @@
 package org.deal.model;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -17,7 +19,7 @@ public class Employment {
     private String status;
 
     private String employerInn;
-    private Double salary;
+    private BigDecimal salary;
     private String position;
     private Integer workExperienceTotal;
     private Integer workExperienceCurrent;
